requirement.txt
asgiref==3.2.3
Django==3.0.3
django-tinymce4-lite==1.7.5
jsmin==2.2.2
pi==0.1.2
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
